Collapse ft params
