+++ GPTifier/src/interface/command_costs.cpp
@@ -1,7 +1,6 @@
 #include "interface/command_costs.hpp"
 
 #include "interface/help_messages.hpp"
-#include "interface/params.hpp"
 #include "serialization/costs.hpp"
 #include "utils.hpp"
